Fix #23: Support external module tests
Modifies ptr's handling of test suites to support test suites that
expect to be in the import path.  Looks for files or directories in the
module directory matching the `test_suite` config value, and also adds a
new config value `test_suite_extras` to list more files or directories
needed for tests. ptr will then copy all of these files and directories
into the root of the virtualenv, and then use the `-m` flag to coverage
to run the test suite. This works for ptr's single-file test suite
(with fixtures and config files listed as "extras") as well as for
package/module-based tests like those in aiosqlite.