Pipe all stderr to stdout
In the case of mypy and linters, their expected outputs are on stdout,
but in the case of fatal errors during execution, they will output on
stderr.  In order to not need to look at both, and to remove the need to
keep a list of steps that output to one or the other, this makes all
steps send stderr to stdout, and always looks at stdout.